payment updated

+++ OnlineFoodDeliverySystem/Serivces/PaymentService.cs
@@ -1,4 +1,5 @@
 ﻿using OnlineFoodDeliverySystem.Data;
+using OnlineFoodDeliverySystem.Exceptions;
 using OnlineFoodDeliverySystem.Models;
 using OnlineFoodDeliverySystem.Repository;
 
@@ -17,26 +18,49 @@ namespace OnlineFoodDeliverySystem.Services
 
         public async Task<IEnumerable<Payment>> GetAllPaymentsAsync()
         {
-            return await _paymentRepository.GetAllPaymentsAsync();
+            var payments = await _paymentRepository.GetAllPaymentsAsync();
+            return payments.ToList();
         }
 
         public async Task<Payment> GetPaymentByIdAsync(int paymentId)
         {
-            return await _paymentRepository.GetPaymentByIdAsync(paymentId);
+            var payment = await _paymentRepository.GetPaymentByIdAsync(paymentId);
+            if (payment == null)
+            {
+                throw new NotFoundException($"Payment with id {paymentId} does not exists");
+            }
+            return payment;
+
         }
 
         public async Task AddPaymentAsync(Payment payment)
         {
+            var paymentExists = await _paymentRepository.GetPaymentByIdAsync(payment.PaymentID);
+            if (paymentExists != null)
+            {
+                throw new AlreadyExistsException($"Payment with id {payment.PaymentID} already exists");
+            }
             await _paymentRepository.AddPaymentAsync(payment);
         }
 
-        public async Task UpdatePaymentAsync(Payment payment)
+        public async Task UpdatePaymentAsync(int id, Payment payment)
         {
-            await _paymentRepository.UpdatePaymentAsync(payment);
+            var paymentExists = await _paymentRepository.GetPaymentByIdAsync(id);
+            if (paymentExists == null)
+            {
+                throw new NotFoundException($"Payment with id {payment.PaymentID} does not exists");
+            }
+            await _paymentRepository.UpdatePaymentAsync(id, payment);
+
         }
 
         public async Task DeletePaymentAsync(int paymentId)
         {
+            var payment = await _paymentRepository.GetPaymentByIdAsync(paymentId);
+            if (payment == null)
+            {
+                throw new NotFoundException($"Payment with id {paymentId} does not exists");
+            }
             await _paymentRepository.DeletePaymentAsync(paymentId);
         }
     }
